recommendation-algorithm-service/requirements.txt
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pymysql==1.1.0
sqlalchemy==2.0.20
scipy==1.11.1
flask-cors==4.0.0
python-dotenv==1.0.0